+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+## 0.17.1
+  * Add simple circuit splitting for SONATA (BRBLD-90).
+
 ## 0.17.0
   * Update ``write_network_config`` to be compatible with the new BBP sonata specs, moved from
     ``brainbuilder.utils.sonata.convert`` to ``brainbuilder.utils.sonata.config`` (NSETM-1526).